pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssiamb"
version = "0.0.4"
description = "SSI Ambiguous Site Detection Tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    {name = "SSI Denmark", email = "support@ssi.dk"},
]
keywords = [
    "bioinformatics",
    "genomics", 
    "variant-calling",
    "mapping",
    "bacterial-genomics"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pysam>=0.21.0",
    "biopython>=1.81",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/ssi-dk/ssiamb"
Repository = "https://github.com/ssi-dk/ssiamb"
Issues = "https://github.com/ssi-dk/ssiamb/issues"

[project.scripts]
ssiamb = "ssiamb.cli:app"

[tool.hatch.version]
path = "src/ssiamb/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/ssiamb"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "pytest --cov=ssiamb --cov-report=term-missing {args:tests}"
cov = "test-cov"

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.hatch.envs.lint.scripts]
check = [
    "ruff check .",
    "black --check --diff .",
    "mypy src/ssiamb",
]
format = [
    "ruff check --fix .",
    "black .",
]

[tool.hatch.envs.docs]
detached = true
dependencies = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-autorefs>=0.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build --clean"
serve = "mkdocs serve --dev-addr localhost:8000"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
